% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryOpenAlex.R
\name{extractAuthors}
\alias{extractAuthors}
\title{Takes set of results from query, turns author info into 1-line-per-result dt}
\usage{
extractAuthors(x)
}
\arguments{
\item{x}{a parsed json list obtained via openAlex query}
}
\description{
Takes set of results from query, turns author info into 1-line-per-result dt
}
