% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryOpenAlex.R
\name{extractBiblio}
\alias{extractBiblio}
\title{Takes set of results from query, turns bibliometric info into 1-line-per-result dt}
\usage{
extractBiblio(x)
}
\arguments{
\item{x}{a parsed json list obtained via openAlex query}
}
\description{
Takes set of results from query, turns bibliometric info into 1-line-per-result dt
}
