% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryOpenAlex.R
\name{extractSource}
\alias{extractSource}
\title{Takes set of results from query, turns source info into 1-line-per-result dt}
\usage{
extractSource(
  x,
  vars = c("id", "display_name", "issn_l", "host_organization", "type")
)
}
\arguments{
\item{x}{a parsed json list obtained via openAlex query}

\item{vars}{variable names to pull from sublist}
}
\description{
Takes set of results from query, turns source info into 1-line-per-result dt
}
