% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractVenues.R
\name{extractVenues}
\alias{extractVenues}
\title{Extract works associated with a venue in openAlex}
\usage{
extractVenues(
  mailto = NULL,
  venue_id = NULL,
  venue_page = NULL,
  cursor = T,
  per_page = NULL,
  to_date = NULL,
  from_date = NULL,
  keep_paratext = FALSE,
  debug = FALSE,
  sleep_time = 0.1
)
}
\arguments{
\item{mailto}{email address of user, needed to get in 'polite pool' of API}

\item{venue_id}{string for openAlex venue ID}

\item{venue_page}{string for openAlex venue page url}

\item{cursor}{boolean if TRUE will perform cursor pagination needed for iterating}

\item{per_page}{how many works returned per page}

\item{to_date}{latest publication date, in YYYY-MM-DD format or a YEAR (assumes YEAR/12/31)}

\item{from_date}{earliest publication date, in YYYY-MM-DD format a YEAR (assumes YEAR/01/01)}

\item{keep_paratext}{boolean to retain or exclude paratext from returns}

\item{debug}{boolean, if TRUE returns query url, if FALSE actually does query}

\item{sleep_time}{time to Sys.sleep() in between cursor iterations}
}
\description{
Primary use is to query a venue ID and extract associated works, e.g., all article records from a given journal
}
