% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryOpenAlex.R
\name{extractWork}
\alias{extractWork}
\title{Takes set of results from query, turns works info into 1-line-per-result dt}
\usage{
extractWork(
  x,
  vars = c("id", "doi", "title", "publication_date", "type", "cited_by_count",
    "is_paratext", "updated_date")
)
}
\arguments{
\item{x}{a parsed json list obtained via openAlex query}

\item{vars}{variable names to pull from sublist}
}
\description{
Takes set of results from query, turns works info into 1-line-per-result dt
}
