% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findconceptID.R
\name{find_concepts}
\alias{find_concepts}
\title{Find concepts and ids for use in openAlex}
\usage{
find_concepts(query_string, which_column = "normalized_name", maxlevel = Inf)
}
\arguments{
\item{query_string}{string or regular expression for searching concept descriptions}
}
\value{
dataframe of results
}
\description{
Search spreadsheet of openAlex concept tree (see https://docs.openalex.org/about-the-data/concept). Current google sheet url is: https://docs.google.com/spreadsheets/d/1LBFHjPt4rj_9r0t0TTAlT68NwOtNH8Z21lBMsJDMoZg/edit#gid=1473310811
}
\details{
NOTE: this function uses the Aug 2022 concept sheet
}
