% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{flattenAuthorNames}
\alias{flattenAuthorNames}
\title{openAlex provides A LOT of author info stored in json. this goes into the nexted structure and pulls out just the author names}
\usage{
flattenAuthorNames(authors_json)
}
\arguments{
\item{authors_json}{a list in json structure that contains openAlex works}
}
\value{
a vector of strings, 1 string for each work, of author names pasted together separated by ";"
}
\description{
openAlex provides A LOT of author info stored in json. this goes into the nexted structure and pulls out just the author names
}
