% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryTitles.R
\name{generateTitleQuery}
\alias{generateTitleQuery}
\title{Take a title string and make a query}
\usage{
generateTitleQuery(
  title = NULL,
  mailto = NULL,
  max_results = 5,
  url = "https://api.openalex.org/works"
)
}
\arguments{
\item{title}{a string name for title https://docs.openalex.org/api}

\item{mailto}{email address of user, needed to get in 'polite pool' of API}

\item{max_results}{how many results to return (uses "per_page" feature on OA to grab first page of results)}

\item{url}{the base url for openAlex query}
}
\description{
Primary use of this function is to make a query url for openAlex
}
