% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json2dt.R
\name{json2dt}
\alias{json2dt}
\title{Read in json.gz file, select data needed for citationSearch, and generate data.table}
\usage{
json2dt(jfile, source = "openAlex", file_location = "workspace")
}
\arguments{
\item{jfile}{a json.gz file (NOTE: easy to make this adaptable, will do...)}

\item{file_location}{can either take a workspace object (location = 'workspace') or a saved json.gz file (location = 'disk'))}
}
\description{
openAlex queries produce tons of data. queryWorks retains a subset for saving as a json.gz, but these data are still way more than is needed for citationSearch. json2dt reads in an openAlex concept search result file, selects necessary metadata items, and generates a data.table object where each OA work is one row, with columns appropriately titled for use as citationSearch index.
}
