% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookupJournal.R
\name{lookupJournal}
\alias{lookupJournal}
\title{Find a journal in openAlex using issn code}
\usage{
lookupJournal(issn = NULL, mailto = NULL, type = NULL)
}
\arguments{
\item{issn}{a string issn value (see https://docs.openalex.org/api-entities/sources/get-a-single-source)}

\item{mailto}{email address of user, needed to get in 'polite pool' of API}
}
\description{
Primary use of this function is to get internal (OpenAlex) venue ID for a journal
}
