% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parseLocationObject}
\alias{parseLocationObject}
\title{openAlex stores information about a work's location(s) as a list nested in the larger works object. This function takes the works object and returns a flatted version of the location object which can be "cbinded" with the works object.}
\usage{
parseLocationObject(work = NULL, primary = TRUE)
}
\arguments{
\item{work}{an openAlex works object}

\item{primary}{a boolean value, default = TRUE, for whether to return \emph{all} locations (FALSE) or the primary location (TRUE) (see https://docs.openalex.org/api-entities/works/work-object/location-object)}
}
\value{
a flattened data.table of location info
}
\description{
openAlex stores information about a work's location(s) as a list nested in the larger works object. This function takes the works object and returns a flatted version of the location object which can be "cbinded" with the works object.
}
