% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryOpenAlex.R
\name{processOA}
\alias{processOA}
\title{Take a json query result from openAlex and parse it into a flat data.frame}
\usage{
processOA(response, keycol = "work.id")
}
\arguments{
\item{response}{a json object obtained from an openAlex query}

\item{keycol}{character string representing column of ID values}
}
\description{
Primary use of this function is to generate a flat file of results to include in search index
}
