% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processWork.R
\name{processWorks}
\alias{processWorks}
\title{Vectorized processing of openalex results}
\usage{
processWorks(
  work = NULL,
  data_style = c("bare_bones", "citation", "comprehensive", "all")
)
}
\arguments{
\item{work}{one or more openalex works entities}

\item{data_style}{options for how much/how little data to return, see @details}
}
\value{
a data.table object with one work per row
}
\description{
takes json works returned by an openAlex query and returns a 1 by X list of citation information in 1 row x Y column data.tables
}
