% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryOpenAlex.R
\name{queryOpenAlex}
\alias{queryOpenAlex}
\title{Search for and return candidate matches to a title in openAlex}
\usage{
queryOpenAlex(
  url = "https://api.openalex.org/works",
  title = NULL,
  mailto = NULL,
  wait_time = 5,
  max_results = 5
)
}
\arguments{
\item{url}{the base url for openAlex query}

\item{title}{a string name for title https://docs.openalex.org/api}

\item{mailto}{email address of user, needed to get in 'polite pool' of API}

\item{wait_time}{how long to wait on query results before timing out, in seconds}

\item{max_results}{how many results to return (uses "per_page" feature on OA to grab first page of results)}
}
\description{
Primary use of this function is to make a query object for use in openAlex that returns potential matches to a title string
}
