% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{streamGZ}
\alias{streamGZ}
\title{simple wrapper to combine stream_in and gzfile to read json.gz OA files}
\usage{
streamGZ(x)
}
\arguments{
\item{x}{a json.gz file}
}
\value{
a decompressed json file
}
\description{
This function just combines two existing functions. Stream_in is preferred over normal read functions because the files are very large.
}
