% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subitems2vector}
\alias{subitems2vector}
\title{this is just a shorthand to go into lists within lists and grab items}
\usage{
subitems2vector(list, var, subvar, fill = NA)
}
\arguments{
\item{list}{a list of json openAlex results}

\item{var}{the name of the variable to pull out from the 2nd level list}

\item{subvar}{the name of the 2nd level list}

\item{fill}{replaces empty values with NA before collapsing to vector}
}
\description{
this is just a shorthand to go into lists within lists and grab items
}
