% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titles.R
\docType{data}
\name{titles}
\alias{titles}
\title{This is a sample of random "titles" (not all are real titles) extracted from a document.}
\description{
This is a sample of random "titles" (not all are real titles) extracted from a document.
}
\keyword{data}
