% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/works2dt.R
\name{venueworks2dt}
\alias{venueworks2dt}
\title{takes a json formatted list of works associated with a host venue and outputs a data.table of simple citation information}
\usage{
venueworks2dt(venueworks_list)
}
\arguments{
\item{venueworks_list}{a list output by calling extractVenues.R}
}
\value{
a data.table object with one work per row
}
\description{
takes a json formatted list of works associated with a host venue and outputs a data.table of simple citation information
}
