% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractWorks.R
\name{extractWorks}
\alias{extractWorks}
\title{Extract works associated with a concept in openAlex}
\usage{
extractWorks(
  dest_file = NULL,
  mailto = NULL,
  concept_id = NULL,
  concept_page = NULL,
  venue_id = NULL,
  venue_page = NULL,
  cursor = T,
  per_page = NULL,
  to_date = NULL,
  from_date = NULL,
  keep_paratext = FALSE,
  debug = FALSE,
  sleep_time = 0.1,
  reduce = T,
  return = T
)
}
\arguments{
\item{dest_file}{location to save output as a json.gz}

\item{mailto}{email address of user, needed to get in 'polite pool' of API}

\item{concept_id}{string for openAlex concept ID}

\item{concept_page}{string for openAlex concept page url}

\item{cursor}{boolean if TRUE will perform cursor pagination needed for iterating}

\item{per_page}{how many works returned per page}

\item{to_date}{latest publication date, in YYYY-MM-DD format or a YEAR (assumes YEAR/12/31)}

\item{from_date}{earliest publication date, in YYYY-MM-DD format a YEAR (assumes YEAR/01/01)}

\item{keep_paratext}{boolean to retain or exclude paratext from returns}

\item{debug}{boolean, if TRUE returns query url, if FALSE actually does query}

\item{sleep_time}{time to Sys.sleep() in between cursor iterations}

\item{reduce}{boolean for whether to reduce scope of final results, see @details}

\item{return}{boolean for whether final result should be returned as workspace object}
}
\description{
Primary use is to query a concept ID and extract associated works, e.g., all article records associated with "habitat"
}
\details{
Note that because extracted records can be pretty large, there is an optional "reduce" command that selects out a subset of key variables before saving or returning the final json list
}
