% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parseGrantsObject}
\alias{parseGrantsObject}
\title{openAlex stores information about a work's funding as a list nested in the larger works object. This function takes the works object and returns a flatted version of the location object which can be "cbinded" with the works object.}
\usage{
parseGrantsObject(work = NULL)
}
\arguments{
\item{work}{an openAlex works object}
}
\value{
a flattened data.table of funding info
}
\description{
openAlex stores information about a work's funding as a list nested in the larger works object. This function takes the works object and returns a flatted version of the location object which can be "cbinded" with the works object.
}
