% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processWork.R
\name{processWork}
\alias{processWork}
\title{takes a json formatted list of works associated with a host venue and outputs a data.table of simple citation information}
\usage{
processWork(work = NULL, style = c("bare_bones", "citation", "custom", "all"))
}
\arguments{
\item{a}{list output generated by extractWorks.R}

\item{data_style}{options for how much/how little data to return, see @details}
}
\value{
a data.table object with one work per row
}
\description{
takes a list of openAlex works and returns a data.table of simple citation information
}
\details{
Note that because extracted records can be pretty large--and are complicated, nested json file--there is an optional "data_style" command that lets the user specify what to return. Currently there are three options: (1) bare_bones returns OpenAlex ID + DOI, basically, results that can be used to look up the work again; (2) citation returns typical citation information, like journal name, author, etc., with a couple bonus items like source.id to link back to openAlex (3) custom can be tailored to suit researcher aims (note this likely requires building a funciton or two to parse a nested list of interest); and (4) all returns the entire result in original json format.
}
