% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processWork.R
\name{processWorks}
\alias{processWorks}
\title{Vectorized processing of work json data}
\usage{
processWorks(
  work = NULL,
  data_style = c("bare_bones", "citation", "comprehensive", "all")
)
}
\arguments{
\item{work}{a list of works results}

\item{data_style}{= c('bare_bones','citation','comprehensive','all')}
}
\description{
Same function but can take a list of works. Not advisable for large data sets, parallelize queryOpenAlex instead.
}
