% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querySources.R
\name{querySources}
\alias{querySources}
\title{Find sources (journals) and source ids for use in openAlex}
\usage{
querySources(source = NULL, mailto = NULL, type = NULL)
}
\arguments{
\item{source}{a string source name https://docs.openalex.org/api}

\item{mailto}{email address of user, needed to get in 'polite pool' of API}

\item{type}{which type of source should be included in query, defaults to all}
}
\description{
Primary use of this function is to get source ID for use in API
}
