% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryTitle.R
\name{queryTitles}
\alias{queryTitles}
\title{Vectorized search for and return candidate matches to a title in openAlex}
\usage{
queryTitles(
  title = NULL,
  mailto = NULL,
  wait_time = 5,
  max_results = 5,
  url = "https://api.openalex.org/works",
  data_style = c("bare_bones", "citation", "comprehensive", "all")
)
}
\arguments{
\item{title}{a vector of title strings https://docs.openalex.org/api}

\item{mailto}{email address of user, needed to get in 'polite pool' of API}

\item{wait_time}{how long to wait on query results before timing out, in seconds}

\item{max_results}{how many results to return (uses "per_page" feature on OA to grab first page of results)}

\item{url}{the base url for openAlex query}
}
\description{
Same function but can take a set of titles. Not advisable for large data sets, parallelize queryOpenAlex instead.
}
