% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryVenues.R
\name{queryVenues}
\alias{queryVenues}
\title{Query openAlex for information about a venue (i.e., journal)}
\usage{
queryVenues(mailto = NULL, venue_string = NULL)
}
\arguments{
\item{mailto}{email address of user, needed to get in 'polite pool' of API}

\item{venue}{a string venue name https://docs.openalex.org/api}
}
\description{
Primary use of this function is to get venue ID for use in API
}
